+++ Data/Entities/User/Ingredient.cs
@@ -36,11 +36,11 @@ public class Ingredient
     public bool SkipShoppingList { get; set; }
 
     [Display(Name = "Serving Size (grams)")]
-    public int ServingSizeGrams { get; set; }
+    public double ServingSizeGrams { get; set; }
 
     public double CaloriesPerServing { get; set; }
 
-    public int GramsPerCup { get; set; }
+    public double GramsPerCup { get; set; }
 
     /// <summary>
     /// Notes about the variation (externally shown).
